Fixed a bunch of conversion errors

+++ include/srtc/rtp_packet_source.h
@@ -24,7 +24,7 @@ public:
 private:
     const uint32_t mSSRC;
     const uint8_t mPayloadId;
-    RandomGenerator<uint32_t> mRandom;
+    RandomGenerator<uint16_t> mRandom;
     uint32_t mGeneratedCount;
     uint32_t mRollover;
     uint16_t mNextSequence;
